requirements.txt
anthropic>=0.40.0
gtts>=2.5.1
moviepy==1.0.3
Pillow>=10.0.0
python-dotenv>=1.0.0
notion-client>=2.2.1
pydub>=0.25.1
